requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.1.0
pydantic[email]>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=0.1.0
python-multipart>=0.0.12
openai>=1.3.0
